feat: split ilab config init
split `ilab config init` into two packages `src/instructlab/cli/config/init.py` and `src/instructlab/config/init.py`

`initialize_config` is the new method in the `config/init.py` package and acts as a standalone way to initialize an instructlab config file via python.

This method needs to take a few arguments since it can be "driven" from the CLI or via someone calling the method manually. Most args are meant to be passed via the CLI and can left blank
